% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission.R
\name{add_commission}
\alias{add_commission}
\title{Add IB's stock commission of fixed plan.}
\usage{
add_commission(data, comm_fun = calc_ib_commission_fixed)
}
\arguments{
\item{data}{data.frame contains "entry" and "exit" price and "qty"}

\item{comm_fun}{commission functions whose args are "qty" and "price"}
}
\value{
data.frame
}
\description{
Add IB's stock commission of fixed plan.
}
