% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction.R
\name{add_pnl}
\alias{add_pnl}
\title{Add cost and Pnl}
\usage{
add_pnl(trades)
}
\arguments{
\item{trades}{data.frame of trades contains "entry" and "exit" price and qty}
}
\value{
data.frame
}
\description{
Add cost and Pnl
}
