% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{boxplot_performance}
\alias{boxplot_performance}
\title{Build boxplot}
\usage{
boxplot_performance(df, x, y)
}
\arguments{
\item{df}{data.frame}

\item{x}{X-axis parameter}

\item{y}{Y-axis parameter}
}
\value{
ggplot
}
\description{
Build boxplot
}
