% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commission.R
\name{calc_ib_commission_fixed}
\alias{calc_ib_commission_fixed}
\title{Calculate IB's stock commission of fixed plan.}
\usage{
calc_ib_commission_fixed(qty, price)
}
\arguments{
\item{qty}{Quantity}

\item{price}{Price}
}
\value{
Numeric scalar
}
\description{
Calculate IB's stock commission of fixed plan.
}
