% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calc_min_ret_performance}
\alias{calc_min_ret_performance}
\title{Calculate minimum performance based on portfolio return}
\usage{
calc_min_ret_performance(ec, clean = FALSE)
}
\arguments{
\item{ec}{Equity curve}

\item{clean}{Whether clean return by clean.boudt}
}
\value{
data.frame
}
\description{
Calculate minimum performance based on portfolio return
}
