% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calc_performance}
\alias{calc_performance}
\title{Calculate performance metrics from equity curve.}
\usage{
calc_performance(ec)
}
\arguments{
\item{ec}{Equity curve}
}
\value{
data.frame
}
\description{
Calculate performance metrics from equity curve.
}
