% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calc_performances}
\alias{calc_performances}
\title{Parallely calculate performance from equity curve}
\usage{
calc_performances(ec, parallel = TRUE)
}
\arguments{
\item{ec}{Equity curve}

\item{parallel}{Whether to calc parallely}
}
\value{
data.frame
}
\description{
Parallely calculate performance from equity curve
}
