% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equity_curve.R
\name{generate_equity_curve}
\alias{generate_equity_curve}
\title{Generate equity curve from transactions.}
\usage{
generate_equity_curve(trans, benchmark = "SPY")
}
\arguments{
\item{trans}{Transactions which contains date and pnl.}

\item{benchmark}{Benchmark symbol.}
}
\value{
data.frame
}
\description{
Generate equity curve from transactions.
}
