% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{generate_params}
\alias{generate_params}
\title{Generate strategy parameters}
\usage{
generate_params(..., constraints = NULL)
}
\arguments{
\item{...}{Vector of parameters}

\item{constraints}{Constraints of perameters}
}
\value{
data.frame
}
\description{
Generate strategy parameters
}
