% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_all_best_params_by_median}
\alias{get_all_best_params_by_median}
\title{Get all best parameter values based on median performance}
\usage{
get_all_best_params_by_median(df, perf_col, select_fun = min)
}
\arguments{
\item{df}{data.frame of performance}

\item{perf_col}{Performance column}

\item{select_fun}{Select function if there are multiple results}
}
\value{
list
}
\description{
Get all best parameter values based on median performance
}
