% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{get_best_param_by_median}
\alias{get_best_param_by_median}
\title{Get best parameter value based on median performance}
\usage{
get_best_param_by_median(df, param_col, perf_col, select_fun = min)
}
\arguments{
\item{df}{data.frame of performance}

\item{param_col}{Parameter column}

\item{perf_col}{Performance column}

\item{select_fun}{Select function if there are multiple results}
}
\value{
scalar
}
\description{
Get best parameter value based on median performance
}
