% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_days.R
\name{get_previous_trading_day}
\alias{get_previous_trading_day}
\title{Get previous trading date of specified date}
\usage{
get_previous_trading_day(end_date)
}
\arguments{
\item{end_date}{Date or character of end date by "yyyy-mm-dd" format}
}
\value{
date
}
\description{
Get previous trading date of specified date
}
