% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_days.R
\name{get_ranges_by_type}
\alias{get_ranges_by_type}
\title{Get trading range by type}
\usage{
get_ranges_by_type(type, start_date, end_date, expand_min_period = 10)
}
\arguments{
\item{type}{Range type like "ExpandingDaily", "Rolling3mDaily" etc.}

\item{start_date}{Date or character of start date by "yyyy-mm-dd" format}

\item{end_date}{Date or character of end date by "yyyy-mm-dd" format}

\item{expand_min_period}{Minimum period for expading type}
}
\value{
data.frame of start and end date
}
\description{
Get trading range by type
}
