% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_days.R
\name{get_trading_days}
\alias{get_trading_days}
\title{Get all trading days}
\usage{
get_trading_days(start_date, end_date)
}
\arguments{
\item{start_date}{Start date by "yyyy-mm-dd" format}

\item{end_date}{End date by "yyyy-mm-dd" format}
}
\value{
Date vector
}
\description{
Get all trading days
}
