% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_equity_curve}
\alias{plot_equity_curve}
\title{Build equity curve and drawdown plot.}
\usage{
plot_equity_curve(ec)
}
\arguments{
\item{ec}{Equity curve}
}
\value{
ggplot
}
\description{
Build equity curve and drawdown plot.
}
