% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_return_histogram}
\alias{plot_return_histogram}
\title{Build histogram of returns from equity curve}
\usage{
plot_return_histogram(ec, binwidth = 0.002)
}
\arguments{
\item{ec}{Equity curve}

\item{binwidth}{Bin width}
}
\value{
ggplot
}
\description{
Build histogram of returns from equity curve
}
