% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{print_performance}
\alias{print_performance}
\title{Format performance for printing purpose}
\usage{
print_performance(ec, col_name, clean = FALSE)
}
\arguments{
\item{ec}{Equity curve}

\item{col_name}{Column name}

\item{clean}{Whether to clean returns by clean.boudt}
}
\value{
data.frame
}
\description{
Format performance for printing purpose
}
