% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{select_param_cols}
\alias{select_param_cols}
\title{Select parameter columns from performance data.frame}
\usage{
select_param_cols(df)
}
\arguments{
\item{df}{data.frame of performance}
}
\value{
data.frame
}
\description{
Select parameter columns from performance data.frame
}
