% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_days.R
\name{get_previous_trading_day}
\alias{get_previous_trading_day}
\title{Get previous trading date of specified date}
\usage{
get_previous_trading_day(date)
}
\arguments{
\item{date}{Date}
}
\value{
date
}
\description{
Get previous trading date of specified date
}
