% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_days.R
\name{get_trading_days}
\alias{get_trading_days}
\title{Get all trading days by SPY and IBM}
\usage{
get_trading_days(start_date = "1993-01-29", end_date = Sys.Date())
}
\arguments{
\item{start_date}{Date or character of start date by "yyyy-mm-dd" format}

\item{end_date}{Date or character of end date by "yyyy-mm-dd" format}
}
\value{
Date vector
}
\description{
Get all trading days by SPY and IBM
}
